requirements.txt
streamlit==1.32.0
Pillow==10.2.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0
torchvision==0.17.0
torchxrayvision==0.0.31
numpy==1.26.3
pandas==2.2.0 